// Архив/src/taxi-bot/scenes/scenes.type.ts
export enum ScenesType {
	RegistrationDriver = 'registrationDriverScene',
	RegistrationPassenger = 'registrationPassengerScene',
	AddAddress = 'addAddressScene',
	DeleteAddress = 'deleteAddressScene',
	EditName = 'editName',
	EditPhone = 'editPhone',
	EditCity = 'editCity',
	CreateOrder = 'createOrder',
	EditNameDriver = 'editNameDriver',
	EditPhoneDriver = 'editPhoneDriver',
	EditCityDriver = 'editCityDriver',
	EditCarDriver = 'editCarDriver',
	EditAccessOrderTypeDriver = 'editAccessOrderTypeDriver',
	BargainOrderByDriver = 'bargainOrderByDriver',
	AccessOrderByDriver = 'accessOrderByDriver',
	AddReview = 'addReview',
}